Code push
I am pushing all the code I have. The code is in shambles and contains some data preprocessor class and then the initial attempt at getting RNN to work.

It currently does not work, as I have no idea how to structure the data, and don't really understand how Keras or Tensorflow works. 

I am in the process of reading a book about it, and plan to continue with this as soon as I understand.